Move utility into its own file
